fix lint err

+++ x/assets/keeper/client_chain.go
@@ -71,8 +71,8 @@ func (k Keeper) GetAllClientChainInfo(ctx sdk.Context) (infos []assetstype.Clien
 	return ret, nil
 }
 
-func (k Keeper) GetAllClientChainID(ctx sdk.Context) ([]uint64, error) {
-	ret := make([]uint64, 0)
+func (k Keeper) GetAllClientChainID(ctx sdk.Context) ([]uint32, error) {
+	ret := make([]uint32, 0)
 	opFunc := func(clientChain *assetstype.ClientChainInfo) error {
 		// #nosec G115
 		ret = append(ret, uint32(clientChain.LayerZeroChainID))
